requirements.txt
fastapi==0.115.4
uvicorn==0.32.0
pydantic==2.10.2
jinja2==3.1.4
python-multipart==0.0.12
websockets==13.1
assemblyai==0.43.1
google-generativeai==0.8.3
requests==2.32.3
aiohttp==3.11.7
newsapi-python==0.2.7

